trying it where i dont resize the window

+++ src/Tile.h
@@ -13,7 +13,7 @@
 class NeighborInfo{
 public:
     int idNum;
-    int freq;
+    float freq;
 };
 
 class Tile{
